Move system APIs to kernel module
