+++ scrypto/src/abi/mod.rs
@@ -1,5 +1,3 @@
 mod component;
-mod kernel;
 
 pub use component::*;
-pub use kernel::*;
