+++ scrypto/src/constructs/account.rs
@@ -1,9 +1,8 @@
 use serde::{Deserialize, Serialize};
 
-use crate::abi::*;
 use crate::constructs::*;
+use crate::kernel::*;
 use crate::types::*;
-use crate::utils::*;
 
 /// An account holds tokens and badges.
 #[derive(Debug, Clone, Serialize, Deserialize)]
