+++ scrypto/src/constructs/badges.rs
@@ -1,9 +1,8 @@
 use serde::{Deserialize, Serialize};
 
-use crate::abi::*;
 use crate::constructs::*;
+use crate::kernel::*;
 use crate::types::*;
-use crate::utils::*;
 
 /// A bucket that holds badge resource.
 #[derive(Debug, Serialize, Deserialize)]
