+++ scrypto/src/constructs/context.rs
@@ -1,6 +1,5 @@
-use crate::abi::*;
+use crate::kernel::*;
 use crate::types::*;
-use crate::utils::*;
 
 /// The execution context.
 pub struct Context {}
