+++ scrypto/src/constructs/logger.rs
@@ -1,5 +1,4 @@
-use crate::abi::*;
-use crate::utils::*;
+use crate::kernel::*;
 
 extern crate alloc;
 use alloc::string::String;
