+++ scrypto/src/constructs/tokens.rs
@@ -1,7 +1,6 @@
-use crate::abi::*;
 use crate::constructs::*;
+use crate::kernel::*;
 use crate::types::*;
-use crate::utils::*;
 
 use serde::{Deserialize, Serialize};
