+++ scrypto/src/constructs/tokens_ref.rs
@@ -1,9 +1,8 @@
 use serde::{Deserialize, Serialize};
 
-use crate::abi::*;
 use crate::constructs::*;
+use crate::kernel::*;
 use crate::types::*;
-use crate::utils::*;
 
 /// A borrowed reference to a `Tokens` bucket.
 #[derive(Debug, Serialize, Deserialize)]
