+++ scrypto/src/utils/mod.rs
@@ -1,7 +1,5 @@
 mod hex;
 mod sha;
-mod syscall;
 
 pub use hex::*;
 pub use sha::*;
-pub use syscall::*;
